Time: 0 ms (100.00%) | Memory: 9.2 MB (56.08%) - LeetSync

+++ 8-string-to-integer-atoi/string-to-integer-atoi.cpp
@@ -5,10 +5,11 @@ public:
         bool isNeg = false;
         int n = s.size();
         int i =0;
-        long long number = 0;
+        int number =0;
 
 
-        while(i <n && s[i] == ' '){
+
+        while(i<n && s[i] == ' '){
             i++;
         }
 
@@ -26,13 +27,13 @@ public:
         for(; i<n; i++){
             if(s[i] >= '0' && s[i] <= '9'){
                 int digit = s[i] - '0';
-                  
-                  if(number > (INT_MAX - digit) /10){
+
+
+                if(number > (INT_MAX - digit) / 10 ){
                     return isNeg ? INT_MIN : INT_MAX;
-                  }
+                }
 
-                  number = number * 10 + digit;
-
+                number = number * 10 + digit;
             }
 
             else {
@@ -40,7 +41,6 @@ public:
             }
         }
 
-
-        return isNeg ? -number : number;
+        return  isNeg ? -number :number;
     }
 };
